Renamed users route to 'user', more routing
