+++ views/layouts/user.handlebars
@@ -8,7 +8,7 @@
         <link rel="icon" type="image/x-icon" href="/public/images/favicon.ico">
 
         <link rel="stylesheet" href="/public/css/main.css">
-        <link rel="stylesheet" href="/public/css/users.css">
+        <link rel="stylesheet" href="/public/css/user.css">
     </head>
 
     <body>
